backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
cloudinary
qrcode
pillow
python-dotenv
psycopg2-binary
pydantic-settings
psycopg2
bcrypt
sqlalchemy[asyncio] 
asyncpg